AC-415: for easier find-as-you-type
